+++ frontend/src/component/common/StrategySeparator/LegacyStrategySeparator.tsx
