+++ frontend/src/component/playground/Playground/PlaygroundResultsTable/FeatureResultInfoPopoverCell/FeatureStrategyList/StrategyList/StrategyItem/StrategyExecution/CustomStrategyParams/CustomStrategyParams.tsx
@@ -5,7 +5,7 @@ import {
     parseParameterStrings,
 } from 'utils/parseParameter';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import { StrategySeparator } from 'component/common/StrategySeparator/StrategySeparator';
+import { StrategySeparator } from 'component/common/StrategySeparator/LegacyStrategySeparator';
 import { useStrategies } from 'hooks/api/getters/useStrategies/useStrategies';
 import { CustomParameterItem } from './CustomParameterItem/CustomParameterItem';
